+++ src/Executor/JobExecutor.php
@@ -3,7 +3,9 @@
 namespace Orisai\Scheduler\Executor;
 
 use DateTimeImmutable;
+use Generator;
 use Orisai\Scheduler\Exception\RunFailure;
+use Orisai\Scheduler\Status\JobSummary;
 use Orisai\Scheduler\Status\RunSummary;
 
 interface JobExecutor
@@ -11,8 +13,9 @@ interface JobExecutor
 
 	/**
 	 * @param list<int|string> $ids
+	 * @return Generator<int, JobSummary, void, RunSummary>
 	 * @throws RunFailure
 	 */
-	public function runJobs(array $ids, DateTimeImmutable $runStart): RunSummary;
+	public function runJobs(array $ids, DateTimeImmutable $runStart): Generator;
 
 }